+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 87cf90f - v7
+| Author: FJrodafo - Sat, 27 Jul 2024 13:55:57 +0100
+| 
 * Commit: c968422 - v6
 | Author: FJrodafo - Wed, 24 Jul 2024 22:27:52 +0100
 |